test: Add Test type
The test.Test type implement the basic operations for a DR test:

- Deploy
- Undeploy
- Protect
- Unprotect
- Failover
- Relocate

We will use it to run a complete DR test flow.

The operations log a progress message, call the e2e code implementing
the operation, and log a completion message.

When a test operation fails the test is failed and cannot continue. We
log the error to the test log, and return an error to the caller to
signal the failure. The caller must abort the test and report the error
to the console.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>