Separate tests from setup and cleanup
This matches better the command behavior.

For test run:

1. Run setup
2. Run tests

For test clean:

1. Clean tests
2. Run cleanup

The natural place for running or cleaning a test seems to be the test
command. already used to run the setup and cleanup steps. This will also
make it easier to integrate the test report in the next step.

This makes is easier to handle test failure and support running multiple
tests in parallel.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>