ci: Run tests on all platforms and architectures
We want to ensure that ramenctl works on all supported platforms. The
best way to do this is to run the unit tests.

Add test matrix job running the tests on all available github runners[1]:

- ubuntu-24.04 (x86_64)
- ubuntu-24.04-arm (arm64)
- windwos-latest (x86_64)
- macos-13 (x86_64)
- macos-15 (arm64)

We cannot use the build job since all jobs run on ubuntu-latest (x86_64)
and we use cross compiling to build the executables for all os and
architectures.

[1] https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories

Fixes: #42
Signed-off-by: Nir Soffer <nsoffer@redhat.com>