test: Separate console.Pass messages
When test passes we want to a user friendly completion message to the
console. Currently we format the message in the test function and pass
it to pass() which is not helpful.

When a test fails we format the failure message in the test function,
but when a test is canceled we need to detect cancellation and log a
different console error message, so this should be done inside fail().

Signed-off-by: Nir Soffer <nsoffer@redhat.com>