test: Simplify step handling and logging
Add Test.start() helper creating a step and adding it to the steps list.
This can make it easier to debug since the state of the tests is
reflected in the steps list.

Since we know the current step, we don't need to pass the step the
fail() and pass().

Add logs when test is started and passed, so we have the complete flow
of a test in ramenctl level.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>