test: Setup, validate and cleanup cancellation
Same changes added to Test for Command, detecting if a step was canceled
and marking the canceled step as canceled. Adding the step to the report
mark the report as canceled.

Similar to tests we have startStep(), failStep() and passStep(). We keep
the current step in the Command.Current and add it to the report when
the stop completes.

Similar to tests we step events.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>