test: Make command easier to test
Testing test.Command requires a real config file for a real cluster.
When we create command.Command we load the config and create an Env,
fetching the clusterClaim to get the managed clusters name.

For fast tests we would like to use test config and env without a real
cluster. Make this possible by changing test.newCommand() to accept a
*command.Command instead of its arguments.

More work is needed to create a command.Command for testing.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>