command: Add command.ForTest
ForTest creates a command configured for tests. The command accept
existing conig and env and does not handle signals.

Running the tests on windows revealed that we are leaking the log file
descriptor when the command completes. This is not an issue in real
usage since we always exit, but in the test it break the test temporary
directory removal:

    testing.go:1232: TempDir RemoveAll cleanup: remove
    C:\Users\RUNNER~1\AppData\Local\Temp\TestRunDisappFailed2850929852\001\test-run.log:
    The process cannot access the file because it is being used by
    another process.

To avoid we sync and close the log in Stop(). The function was renamed
to Close() to make it more clear.

We will use this to create fast tests without real clusters.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>