build: Harden workflow permissions
When workflow do no specify permissions, they inherit the permissions
from the repository permissions. Often these permissions do not adhere
to the principle of least privilege and can be reduced to read-only.

The test workflows does not need to change anything in the repository so
it should use read-only permission.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>