report: Move step management to generic report
Decouple Step from Test by replacing Step.AddTest with Step.AddStep.
When we add tests to the "tests" step we create a step from test. With
this change Step is generic and can move up to the generic report
package. Adding sub steps will be used for other commands and can be
used to group setup and cleanup steps.

The Step type and functions managing it moved to the report package. The
test.Report is much simpler now, adding the test config to the report
and populating the test summary.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>