gather: Return results from gather.Namespaces()
Previously gather.Namespaces() handled gather results and logged errors
and console messages. This worked for the test command but will not work
for the gather command.

Change gather.Namespaces() to return a channel for getting gather
results. The caller can iterate of the results to log errors, console
messages or build the report.

The test command was updated to handle logging during gather. We can
update it later to add the gather steps to the report.

Fixes: #181
Signed-off-by: Nir Soffer <nsoffer@redhat.com>