report: Rename Report to Base
We have test.Report and validate.Report embedding report.Report. However
validate.Report is also needed for the gather command, so we want to
move it up the report package. This change rename report.Report to
report.Base and report.New() to report.NewBase(), to make room for the
new common report.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>