validate: Validate drpc.Action
This should never fail but would nice to validate since it is easy to
create a drpc with the wrong action name. The operator should prevent
this but it is possible to scale down the operator and add a drpc with
an unknown action name.

The action value was to change to omitempty so we don't know empty
value.

Example report:

    applicationStatus:
      hub:
        drpc:
          action:
            state: ok ✅
          conditions:
          - state: ok ✅
            type: Available
          ...

Part-of: #262
Signed-off-by: Nir Soffer <nsoffer@redhat.com>