validate: Validate drpc phase
Validating phase is more complicated, since it depends on the action. We
simplify validation by marking the stable phase for every action as OK,
and anything else as an error. The reason is that an application should
not be in unstable phase (e.g. FailingOver) for long time. If it is
stuck at unstable state for long time this is a problem and we need to
investigate.

The unit test validate that we return the expected ValidatedString for
every possible combination of action and phase.

Example report:

    applicationStatus:
      hub:
        drpc:
          action:
            state: ok ✅
          conditions:
          - state: ok ✅
            type: Available
          - state: ok ✅
            type: PeerReady
          - description: VolumeReplicationGroup (openshift-dr-ops/rhel9-snapshot-vm-drpc)
              on cluster prsurve-c2-7j is not reporting any lastGroupSyncTime as primary,
              retrying till status is met
            state: error ❌
            type: Protected
          deleted:
            state: ok ✅
          drPolicy: dr-policy-4m
          name: rhel9-snapshot-vm-drpc
          namespace: openshift-dr-ops
          phase:
            state: ok ✅
            value: Deployed
          progression: Completed
          ...

Part-of: #262
Signed-off-by: Nir Soffer <nsoffer@redhat.com>