validate: Skip protected pvcs on secondary vrg
With volrep applications the protected pvcs becomes stale on the
secondary cluster, causing validation to fail after failover or
relocate.

In volsync protectedPVCs is always empty on the secondary cluster. We
can validate the VolSync.RDSpec instead, but we don't look at it yet.

Example failed run:

    % ramenctl validate application --name disapp-deploy-rbd --namespace ramen-ops -o out/relocate-11
    ⭐ Using config "config.yaml"
    ⭐ Using report "out/relocate-11"

    🔎 Validate config ...
       ✅ Config validated

    🔎 Validate application ...
       ✅ Inspected application
       ✅ Gathered data from cluster "hub"
       ✅ Gathered data from cluster "dr2"
       ✅ Gathered data from cluster "dr1"
       ❌ Problems found during validation

    ❌ validation failed (22 ok, 1 stale, 1 errors)

The issue is the stale protected pvc for a non-existing pvc on the
secondary cluster:

    secondaryCluster:
      name: dr2
      vrg:
        conditions:
        - state: ok ✅
          type: NoClusterDataConflict
        deleted:
          state: ok ✅
        name: disapp-deploy-rbd
        namespace: ramen-ops
        protectedPVCs:
        - conditions:
          - state: ok ✅
            type: DataReady
          - description: Observed generation 4 does not match object generation 5
            state: "stale \U0001F7E0"
            type: ClusterDataProtected
          deleted:
            description: Resource does not exist
            state: error ❌
            value: true
          name: busybox-pvc
          namespace: e2e-disapp-deploy-rbd
          phase:
            state: ""
          replication: volrep
        state:
          state: ok ✅
          value: Secondary

The application is otherwise ok. This seems to be a ramen bug reporting
stale status instead of removing the protected pvcs from the status like
volsync. However we can easily fix this by dropping the protected pvcs
if the vrg is secondary.

We show the protected pvc in other states to we can see them on the
primary cluster when the vrg is progressing towards primary state.

Fixes: #286
Signed-off-by: Nir Soffer <nsoffer@redhat.com>