docs: Document validate application command
Add `docs/validate.md` describe the validate command. The cluster sub
command is omitted for now since it is not complete yet. The document
shows how to run the command and an example report for an application
without any problems.

We will add more details and problem cases in a `validating.md`, similar
to `testing.md`.

Fixes: #280
Signed-off-by: Nir Soffer <nsoffer@redhat.com>