ramen: Add operator names constants
It seems that the name of the ramen operator depends on the way ramen is
installed. We can find the name by searching for deployment with a ramen
specific label, but this is more complicated, and need to know the ramen
specific label. Another option is to gather the olm subscription and get
the name of the ramen deployment from it. This seem too complicated so I
simplified by by hard coding the names as used in OpenShift and in the
upstream testing environment.

Add a TODO comment to discover the values from the cluster. We can
improve this later.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>