validate: Validate ramen configmaps
When validating ramen on the hub and the managed clusters, validated
also the ramen operator configmap and the s3 secrets references by the
configmap.

We don't validate the following properties:
- ramen configmap on the hub and managed clusters are the same (except
  the controller type).
- s3 secrets are the same on the hub and managed clusters.
- s3 secrets have the required fields

These validations requires larger effort to keep all configmaps and
secrets in memory during validation. Lets do this later as a bug fix or
in the next release.

Fixes: #233
Signed-off-by: Nir Soffer <nsoffer@redhat.com>