docs: Remove flow description
The test flow and clean flow sections describe what the system doing
during `ramenctl test run` and `ramenctl test clean`. However users are
confused by the text, thinking that this is instructions for the user.

Most users are interested in the results of the tests, and not in how
the command is implemented, so this text is mostly unhelpful. Removing
it will make the document simpler and easier to focus on the important
parts.

We can document the flow elsewhere, maybe in the code.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>